+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [3.1.2] (Unreleased)
 
+### Added
+
+- Added the ability to run a script/file when launching a project.
+
 ### Changed
 
 - Now you can specify GitHub URLs without `.git` (`https://github.com/me/my-repo`).